// part9/patientor_backend/src/utils.ts
import { z } from 'zod';
import { Gender, HealthCheckRating, Diagnosis } from './types';

const BaseEntrySchema = z.object({
  // id: z.string(),
  description: z.string(),
  date: z.string().date(),
  specialist: z.string(),
  diagnosisCodes: z.array(z.string()).optional()
});

const HealthCheckEntrySchema = BaseEntrySchema.extend({
  type: z.literal("HealthCheck"),
  healthCheckRating: z.nativeEnum(HealthCheckRating)
});

const DischargeSchema = z.object({
  date: z.string().date(),
  criteria: z.string()
});

const HospitalEntrySchema = BaseEntrySchema.extend({
  type: z.literal("Hospital"),
  discharge: DischargeSchema
});

const SickLeaveSchema = z.object({
  startDate: z.string().date(),
  endDate: z.string().date()
});

const OccupationalHealthcareEntrySchema = BaseEntrySchema.extend({
  type: z.literal("OccupationalHealthcare"),
  employerName: z.string(),
  sickLeave: SickLeaveSchema.optional()
});

export const MedicalEntrySchema = z.discriminatedUnion("type", [
  HealthCheckEntrySchema,
  HospitalEntrySchema,
  OccupationalHealthcareEntrySchema,
]);


export const newEntrySchema = z.object({
  name: z.string(),
  dateOfBirth: z.string().date(),
  ssn: z.string(),
  gender: z.nativeEnum(Gender),
  occupation: z.string(),
  entries: z.array(
    MedicalEntrySchema
  )
})


// const toNewPatientEntry = (object: unknown): PatientEntry => {
//   return newEntrySchema.parse(object);
// };

const parseDiagnosisCodes = (object: unknown): Array<Diagnosis["code"]> => {
  if (!object || typeof object !== "object" || !("diagnosisCodes" in object)) {
    return [] as Array<Diagnosis["code"]>;
  }

  return object.diagnosisCodes as Array<Diagnosis["code"]>;
}

export default {
  // toNewPatientEntry,
  parseDiagnosisCodes
}
